Add files via upload
Phoenixx Club NFTs Token that is a base token for nft collection of seven unique fantasy NFTs, each with a limited edition of only 100 copies. The collection includes the following designs: The Fox, The Dog, The Dragon, The Cat, The Bird, The Horses, and The Pony. These NFTs are part of the Phoenixx Club and offer exclusive benefits to members, such as access to private events and merchandise. The Phoenixx NFTs are available for purchase on various NFT marketplaces, and their value may fluctuate based on market demand.